db/requirements.txt
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.20
munch==2.5.0
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
six==1.16.0
tomli==2.0.0
typing_extensions==4.0.1
